refactor(commands): extract query tests to sibling file
